Rename package fully to matrix-sygnal
