+++ matrix_sygnal/exceptions.py
