+++ matrix_sygnal/http.py
@@ -19,7 +19,7 @@ import logging
 import sys
 import time
 import traceback
-from typing import TYPE_CHECKING, Callable, List, Union
+from typing import TYPE_CHECKING, Any, Callable, List, Union
 from uuid import uuid4
 
 from opentracing import Format, Span, logs, tags
@@ -35,15 +35,15 @@ from twisted.web.http import (
 from twisted.web.resource import Resource
 from twisted.web.server import NOT_DONE_YET
 
-from sygnal.exceptions import (
+from matrix_sygnal.exceptions import (
     InvalidNotificationException,
     NotificationDispatchException,
 )
-from sygnal.notifications import Notification, NotificationContext, Pushkin
-from sygnal.utils import NotificationLoggerAdapter, json_decoder
+from matrix_sygnal.notifications import Notification, NotificationContext, Pushkin
+from matrix_sygnal.utils import NotificationLoggerAdapter, json_decoder
 
 if TYPE_CHECKING:
-    from sygnal.sygnal import Sygnal
+    from matrix_sygnal.sygnal import Sygnal
 
 logger = logging.getLogger(__name__)
 
@@ -177,7 +177,7 @@ class V1NotifyHandler(Resource):
 
             root_span_accounted_for = True
 
-            async def cb():
+            async def cb() -> None:
                 with REQUESTS_IN_FLIGHT_GUAGE.labels(
                     self.__class__.__name__
                 ).track_inprogress():
@@ -342,7 +342,10 @@ class SygnalLoggedSite(server.Site):
     """
 
     def __init__(
-        self, *args, log_formatter: Callable[[str, server.Request], str], **kwargs
+        self,
+        *args: Any,
+        log_formatter: Callable[[str, server.Request], str],
+        **kwargs: Any,
     ):
         super().__init__(*args, **kwargs)
         self.log_formatter = log_formatter
