+++ tests/test_http.py
@@ -16,7 +16,7 @@ from unittest.mock import MagicMock, patch
 
 from aioapns.common import NotificationResult
 
-from sygnal.apnspushkin import ApnsPushkin
+from matrix_sygnal.apnspushkin import ApnsPushkin
 
 from tests import testutils
 
@@ -50,14 +50,16 @@ DEVICE_EXAMPLE_AMBIGIOUS = {
 
 class HttpTestCase(testutils.TestCase):
     def setUp(self) -> None:
-        self.apns_mock_class = patch("sygnal.apnspushkin.APNs").start()
+        self.apns_mock_class = patch("matrix_sygnal.apnspushkin.APNs").start()
         self.apns_mock = MagicMock()
         self.apns_mock_class.return_value = self.apns_mock
 
         # pretend our certificate exists
         patch("os.path.exists", lambda x: x == TEST_CERTFILE_PATH).start()
         # Since no certificate exists, don't try to read it.
-        patch("sygnal.apnspushkin.ApnsPushkin._report_certificate_expiration").start()
+        patch(
+            "matrix_sygnal.apnspushkin.ApnsPushkin._report_certificate_expiration"
+        ).start()
         self.addCleanup(patch.stopall)
 
         super().setUp()
