+++ tox.ini
@@ -18,7 +18,7 @@ extras =
 allowlist_externals = poetry
 
 commands = 
-  poetry run coverage run --source=sygnal -m twisted.trial tests
+  poetry run coverage run --source=matrix_sygnal -m twisted.trial tests
   poetry run coverage report --sort=cover
   poetry run coverage html
 
@@ -27,13 +27,13 @@ commands =
 allowlist_externals = poetry
 
 commands =
-  poetry run ruff sygnal/ tests/ stubs
-  poetry run black --check --diff sygnal/ tests/ stubs
-  poetry run isort --check-only --diff sygnal/ tests/ stubs
+  poetry run ruff matrix_sygnal/ tests/ stubs
+  poetry run black --check --diff matrix_sygnal/ tests/ stubs
+  poetry run isort --check-only --diff matrix_sygnal/ tests/ stubs
 
 [testenv:check_types]
 
 allowlist_externals = poetry
 
 commands =
-  poetry run mypy sygnal/ tests/ stubs
+  poetry run mypy matrix_sygnal/ tests/ stubs
